Update Directory Structure
