+++ src/routes/Comunnity/HotBoard.js
@@ -1,20 +1,20 @@
-import PageHeader from "../components/PageHeader";
+import PageHeader from "../../components/PageHeader";
 import styled from "styled-components";
-import PostCard from "../components/PostCard";
+import PostCard from "../../components/PostCard";
 import Pagination from "@mui/material/Pagination";
 import Stack from "@mui/material/Stack";
-import Notice from "../components/Comunnity/Notice";
+import Notice from "../../components/Comunnity/Notice";
 import { useSelector } from "react-redux";
 
-// 커뮤니티 - 자유게시판
-function FreeBoard() {
+// 커뮤니티 - HOT 게시판
+function HotBoard() {
   // 변수 선언
-  const posts = useSelector((state) => state.posts.free);
+  const posts = useSelector((state) => state.posts.hot);
 
-  // 자유게시판 View
+  // HOT 게시판 View
   return (
-    <FreeBoardContainer>
-      <PageHeader title="자유게시판" subTitle="자유롭게 의견을 나누세요" />
+    <HotBoardContainer>
+      <PageHeader title="H🔥T 게시판" subTitle="떠오르고 있는 인기 글이에요" />
       <Posts>
         <Notice notice="상대방을 비방하는 글은 자제해주세요" />
         {posts.map((post) => (
@@ -30,19 +30,17 @@ function FreeBoard() {
       <Stack className="pagination" spacing={2}>
         <Pagination count={10} color="primary" />
       </Stack>
-    </FreeBoardContainer>
+    </HotBoardContainer>
   );
 }
 
 //////////////////////////////////////// Styled-Components
-const FreeBoardContainer = styled.div`
+const HotBoardContainer = styled.div`
   width: calc(100vw - 160px);
 
   .pagination {
     align-items: center;
-
-    .MuiButtonBase-root {
-    }
+    margin: 40px 0 50px 0;
   }
 `;
 const Posts = styled.div`
@@ -50,4 +48,4 @@ const Posts = styled.div`
   margin: 2vh 72px;
 `;
 
-export default FreeBoard;
+export default HotBoard;
