+++ src/routes/MyInfo/MyPosts.js
@@ -1,10 +1,10 @@
 import { useSelector } from "react-redux";
-import PageHeader from "../components/PageHeader";
+import PageHeader from "../../components/PageHeader";
 import styled from "styled-components";
-import PostCard from "../components/PostCard";
+import PostCard from "../../components/PostCard";
 import Pagination from "@mui/material/Pagination";
 import Stack from "@mui/material/Stack";
-import Notice from "../components/Comunnity/Notice";
+import Notice from "../../components/Comunnity/Notice";
 
 // SideBar - 내 글 보기
 function MyPost() {
